Fix after review.

+++ storage/txcache/concurrentMap.go
@@ -4,22 +4,25 @@ import (
 	"sync"
 )
 
-// ConcurrentMap is a "thread" safe map of type string:Anything.
-// To avoid lock bottlenecks this map is dived to several (ShardCount) map shards.
+// This implementation is a simplified version of:
+// https://github.com/ElrondNetwork/concurrent-map
+
+// ConcurrentMap is a thread safe map of type string:Anything.
+// To avoid lock bottlenecks this map is divided to several (ShardCount) map shards.
 type ConcurrentMap struct {
-	shardCount int
+	shardCount uint32
 	shards     []*ConcurrentMapShard
 }
 
-// ConcurrentMapShard is a "thread" safe string to anything map.
+// ConcurrentMapShard is a thread safe string to anything map.
 type ConcurrentMapShard struct {
-	maxSize      int
-	items        map[string]interface{}
-	sync.RWMutex // Read Write mutex, guards access to internal map.
+	maxSize uint32
+	items   map[string]interface{}
+	sync.RWMutex
 }
 
 // NewConcurrentMap creates a new concurrent map.
-func NewConcurrentMap(maxSize int, shardCount int) *ConcurrentMap {
+func NewConcurrentMap(maxSize uint32, shardCount uint32) *ConcurrentMap {
 	m := ConcurrentMap{
 		shardCount: shardCount,
 		shards:     make([]*ConcurrentMapShard, shardCount),
@@ -32,24 +35,24 @@ func NewConcurrentMap(maxSize int, shardCount int) *ConcurrentMap {
 	if maxSize%shardCount != 0 {
 		shardSize++
 	}
-	for i := 0; i < shardCount; i++ {
+
+	for i := uint32(0); i < shardCount; i++ {
 		m.shards[i] = &ConcurrentMapShard{
 			maxSize: shardSize,
-
-			items: make(map[string]interface{}),
+			items:   make(map[string]interface{}),
 		}
 	}
+
 	return &m
 }
 
 // GetShard returns shard under given key.
 func (m *ConcurrentMap) GetShard(key string) *ConcurrentMapShard {
-	return m.shards[uint(fnv32(key))%uint(m.shardCount)]
+	return m.shards[fnv32(key)%m.shardCount]
 }
 
 // Set sets the given value under the specified key.
 func (m *ConcurrentMap) Set(key string, value interface{}) {
-	// Get map shard.
 	shard := m.GetShard(key)
 	shard.Lock()
 	shard.items[key] = value
@@ -58,10 +61,8 @@ func (m *ConcurrentMap) Set(key string, value interface{}) {
 
 // Get retrieves an element from map under given key.
 func (m *ConcurrentMap) Get(key string) (interface{}, bool) {
-	// Get shard
 	shard := m.GetShard(key)
 	shard.RLock()
-	// Get item from shard.
 	val, ok := shard.items[key]
 	shard.RUnlock()
 	return val, ok
@@ -70,7 +71,7 @@ func (m *ConcurrentMap) Get(key string) (interface{}, bool) {
 // Count returns the number of elements within the map.
 func (m *ConcurrentMap) Count() int {
 	count := 0
-	for i := 0; i < m.shardCount; i++ {
+	for i := uint32(0); i < m.shardCount; i++ {
 		shard := m.shards[i]
 		shard.RLock()
 		count += len(shard.items)
@@ -81,10 +82,8 @@ func (m *ConcurrentMap) Count() int {
 
 // Has looks up an item under specified key.
 func (m *ConcurrentMap) Has(key string) bool {
-	// Get shard
 	shard := m.GetShard(key)
 	shard.RLock()
-	// See if element is within shard.
 	_, ok := shard.items[key]
 	shard.RUnlock()
 	return ok
@@ -92,7 +91,6 @@ func (m *ConcurrentMap) Has(key string) bool {
 
 // Remove removes an element from the map.
 func (m *ConcurrentMap) Remove(key string) {
-	// Try to get shard.
 	shard := m.GetShard(key)
 	shard.Lock()
 	delete(shard.items, key)
@@ -104,14 +102,10 @@ func (m *ConcurrentMap) IsEmpty() bool {
 	return m.Count() == 0
 }
 
-// Iterator callback,called for every key,value found in
-// maps. RLock is held for all calls for a given shard
-// therefore callback sess consistent view of a shard,
-// but not across the shards
+// IterCb is an iterator callback
 type IterCb func(key string, v interface{})
 
-// Callback based iterator, cheapest way to read
-// all elements in a map.
+// IterCb iterates over the map (cheapest way to read all elements in a map)
 func (m *ConcurrentMap) IterCb(fn IterCb) {
 	for idx := range m.shards {
 		shard := (m.shards)[idx]
